# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nextstep Dashboard API
  version: 1.1.0
  description: >-
    Contract for the multi-tenant SML dashboard, production administration,
    distributed report execution, LINE delivery, and LIFF viewer. Mutation
    requests authenticated by cookie also require X-CSRF-Token.
servers:
  - url: https://dashboard.nextstep-soft.com
...
paths:
...
  /api/v1/auth/admin/login:
    post:
      tags: [Admin authentication]
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminLoginInput'}
      responses:
        '200':
          description: Session created and secure cookie set.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/auth/admin/logout:
    post:
      tags: [Admin authentication]
      operationId: logoutAdmin
      security: [{AdminSession: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Session revoked.}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/auth/admin/session:
    get:
      tags: [Admin authentication]
      operationId: getAdminSession
      security: [{AdminSession: []}]
      responses:
        '200':
          description: Current admin session.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/auth/admin/password:
    put:
      tags: [Admin authentication]
      operationId: rotateAdminPassword
      security: [{AdminSession: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminPasswordInput'}
      responses:
        '200':
          description: Password rotated and all other sessions revoked.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/tenants:
    get:
      tags: [Admin tenants]
      operationId: listTenants
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/PageSize'}
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/TenantStatus'}
        - name: search
          in: query
          schema: {type: string, maxLength: 160}
      responses:
        '200':
          description: Cursor-paginated tenants.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      tags: [Admin tenants]
      operationId: createTenant
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TenantInput'}
      responses:
        '201':
          description: Tenant created disabled by default.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Tenant'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: nextstep_admin_session
    ViewerSession:
      type: apiKey
      in: cookie
      name: nextstep_viewer_session
  parameters:
...
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 512}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
    CSRFToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 32, maxLength: 256}}
  responses:
    Unauthorized: {$ref: '#/components/responses/Problem401'}
...
    Conflict: {$ref: '#/components/responses/Problem409'}
...
    ValidationFailed: {$ref: '#/components/responses/Problem422'}
    RateLimited: {$ref: '#/components/responses/Problem429'}
...
    Problem401: {description: Authentication required or invalid., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
    Problem409: {description: 'Duplicate request, state conflict, or optimistic lock conflict.', content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
    Problem422: {description: Input is syntactically valid but violates a business rule., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
    Problem429: {description: Rate or concurrency limit exceeded., headers: {Retry-After: {schema: {type: integer}}}, content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
  schemas:
...
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            retryable: {type: boolean}
            fieldErrors: {type: array, items: {$ref: '#/components/schemas/FieldError'}}
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: {type: string}
        code: {type: string}
        message: {type: string}
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: string, nullable: true}
        hasMore: {type: boolean}
    AdminLoginInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 120}
        password: {type: string, minLength: 1, maxLength: 1024, writeOnly: true}
    AdminSession:
      type: object
      required: [username, expiresAt]
      properties:
        username: {type: string}
        csrfToken: {type: string, description: Returned on login; subsequent requests may read the same-site CSRF cookie.}
        expiresAt: {type: string, format: date-time}
        mustRotateBootstrapPassword: {type: boolean, default: true}
    AdminPasswordInput:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: {type: string, minLength: 1, maxLength: 1024, writeOnly: true}
        newPassword: {type: string, minLength: 15, maxLength: 1024, writeOnly: true}
    TenantStatus: {type: string, enum: [ACTIVE, DISABLED, EXPIRED]}
    Tenant:
      type: object
      required: [id, slug, name, timezone, status, accessEndsAt, version, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        timezone: {type: string, enum: [Asia/Bangkok]}
        status: {$ref: '#/components/schemas/TenantStatus'}
        accessEndsAt: {type: string, format: date-time}
        version: {type: integer, minimum: 1}
        smlReadiness: {type: string, enum: [UNCONFIGURED, UNTESTED, READY, FAILED]}
        nextScheduleAt: {type: string, format: date-time, nullable: true}
        viewerUrl: {type: string, format: uri, readOnly: true, description: Canonical Dashboard link generated from configured PUBLIC_BASE_URL; opening it never grants access.}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    TenantInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        slug: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 80, deprecated: true, description: Compatibility field for older clients; omitted by the current admin UI.}
        name: {type: string, minLength: 1, maxLength: 160}
        timezone: {type: string, enum: [Asia/Bangkok], deprecated: true}
        accessEndsAt: {type: string, format: date-time, deprecated: true}
...
    TenantPage:
      type: object
      required: [data, page]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Tenant'}}
        page: {$ref: '#/components/schemas/PageInfo'}
...